refactor(domain): extract inflection-related classes to model
no behavior change
