+++ src/main/java/com/annepolis/lexiconmeum/shared/model/LexemeBuilder.java
@@ -3,6 +3,8 @@ package com.annepolis.lexiconmeum.shared.model;
 import com.annepolis.lexiconmeum.shared.model.grammar.GrammaticalGender;
 import com.annepolis.lexiconmeum.shared.model.grammar.GrammaticalPosition;
 import com.annepolis.lexiconmeum.shared.model.grammar.InflectionClass;
+import com.annepolis.lexiconmeum.shared.model.inflection.Inflection;
+import com.annepolis.lexiconmeum.shared.model.inflection.InflectionKey;
 
 import java.nio.charset.StandardCharsets;
 import java.util.*;
