+++ src/main/java/com/annepolis/lexiconmeum/shared/model/inflection/BuilderHasGrammaticalCase.java
@@ -1,4 +1,4 @@
-package com.annepolis.lexiconmeum.lexeme.detail;
+package com.annepolis.lexiconmeum.shared.model.inflection;
 
 import com.annepolis.lexiconmeum.shared.model.grammar.GrammaticalCase;
