+++ src/main/java/com/annepolis/lexiconmeum/shared/model/inflection/Declension.java
@@ -1,8 +1,5 @@
-package com.annepolis.lexiconmeum.lexeme.detail.noun;
+package com.annepolis.lexiconmeum.shared.model.inflection;
 
-import com.annepolis.lexiconmeum.lexeme.detail.BuilderHasGrammaticalCase;
-import com.annepolis.lexiconmeum.shared.model.Inflection;
-import com.annepolis.lexiconmeum.shared.model.InflectionBuilder;
 import com.annepolis.lexiconmeum.shared.model.grammar.GrammaticalCase;
 import com.annepolis.lexiconmeum.shared.model.grammar.GrammaticalNumber;
