+++ src/test/java/com/annepolis/lexiconmeum/TestUtil.java
@@ -1,12 +1,12 @@
 package com.annepolis.lexiconmeum;
 
-import com.annepolis.lexiconmeum.lexeme.detail.adjective.Agreement;
-import com.annepolis.lexiconmeum.lexeme.detail.noun.Declension;
-import com.annepolis.lexiconmeum.lexeme.detail.verb.Conjugation;
 import com.annepolis.lexiconmeum.shared.model.Lexeme;
 import com.annepolis.lexiconmeum.shared.model.LexemeBuilder;
 import com.annepolis.lexiconmeum.shared.model.Sense;
 import com.annepolis.lexiconmeum.shared.model.grammar.*;
+import com.annepolis.lexiconmeum.shared.model.inflection.Agreement;
+import com.annepolis.lexiconmeum.shared.model.inflection.Conjugation;
+import com.annepolis.lexiconmeum.shared.model.inflection.Declension;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -135,7 +135,7 @@ public class TestUtil {
                 for (GrammaticalNumber number : GrammaticalNumber.values()) {
                     for (GrammaticalCase grammaticalCase : GrammaticalCase.values()) {
                         String form = generateForm( gender, number, grammaticalCase);
-                        if (form != null) forms.add(form);
+                        forms.add(form);
                     }
                 }
             }
@@ -232,9 +232,7 @@ public class TestUtil {
                 for (GrammaticalNumber number : GrammaticalNumber.values()) {
                     for (GrammaticalCase grammaticalCase : GrammaticalCase.values()) {
                         String form = generateForm( gender, number, grammaticalCase);
-                        if (form != null) {
-                            agreements.add(buildAgreement(form, gender, number, grammaticalCase));
-                        }
+                        agreements.add(buildAgreement(form, gender, number, grammaticalCase));
                     }
                 }
             }
